+++ api/src/reportcreator_api/conf/settings.py
@@ -551,9 +551,9 @@ CELERY_BROKER_CONNECTION_RETRY_ON_STARTUP = True
 # Execute tasks locally, if no broker is configured
 CELERY_TASK_ALWAYS_EAGER = not CELERY_BROKER_URL
 
-# Time limits are only enforced if a broker is configured and an external worker is used (but not in eager mode).
-# Self-hosted SysReptor instances use the eager mode by default, resulting in no PDF rendering time limits being applied.
-PDF_RENDERING_TIME_LIMIT = config('PDF_RENDERING_TIME_LIMIT', cast=int, default=60)
+# Maximum time a PDF rendering task is allowed to run. If a task takes longer, it gets cancelled.
+# Set to 0 to disable the time limit
+PDF_RENDERING_TIME_LIMIT = config('PDF_RENDERING_TIME_LIMIT', cast=int, default=5 * 60)
 
 
 # History